Moved cuckoo_clean to library
