+++ lib/cuckoo/core/startup.py
@@ -3,6 +3,7 @@
 # See the file 'docs/LICENSE' for copying permission.
 
 import os
+import shutil
 import sys
 import copy
 import json
@@ -261,3 +262,63 @@ def init_yara():
             log.debug("\t `-- %s", entry)
         else:
             log.debug("\t |-- %s", entry)
+
+
+def cuckoo_clean():
+    """Clean up cuckoo setup.
+    It deletes logs, all stored data from file system and configured databases (SQL
+    and MongoDB.
+    """
+    # Init logging.
+    # This need to init a console logger handler, because the standard
+    # logger (init_logging()) logs to a file which will be deleted.
+    create_structure()
+    init_console_logging()
+
+    # Initialize the database connection.
+    db = Database()
+
+    # Drop all tables.
+    db.drop()
+
+    # Check if MongoDB reporting is enabled and drop that if it is.
+    cfg = Config("reporting")
+    if cfg.mongodb and cfg.mongodb.enabled:
+        from pymongo import MongoClient
+        host = cfg.mongodb.get("host", "127.0.0.1")
+        port = cfg.mongodb.get("port", 27017)
+        mdb = cfg.mongodb.get("db", "cuckoo")
+        try:
+            conn = MongoClient(host, port)
+            conn.drop_database(mdb)
+            conn.disconnect()
+        except:
+            log.warning("Unable to drop MongoDB database: %s", mdb)
+
+    # Paths to clean.
+    paths = [
+        os.path.join(CUCKOO_ROOT, "db"),
+        os.path.join(CUCKOO_ROOT, "log"),
+        os.path.join(CUCKOO_ROOT, "storage"),
+    ]
+
+    # Delete various directories.
+    for path in paths:
+        if os.path.isdir(path):
+            try:
+                shutil.rmtree(path)
+            except (IOError, OSError) as e:
+                log.warning("Error removing directory %s: %s", path, e)
+
+    # Delete all compiled Python objects ("*.pyc").
+    for dirpath, dirnames, filenames in os.walk(CUCKOO_ROOT):
+        for fname in filenames:
+            if not fname.endswith(".pyc"):
+                continue
+
+            path = os.path.join(CUCKOO_ROOT, dirpath, fname)
+
+            try:
+                os.unlink(path)
+            except (IOError, OSError) as e:
+                log.warning("Error removing file %s: %s", path, e)
